// clients/app/src/utils/tokens/xtc/interfaces.ts
export type BurnError = { InsufficientBalance: null } | { InvalidTokenContract: null } | { NotSufficientLiquidity: null };
export type BurnResult = { Ok: TransactionId } | { Err: BurnError };
// ...
export type TransactionId = bigint;